# openapi-copy.json
{"openapi":"3.1.0","info":{"title":"AI Voice Clone Platform","version":"0.1.0"},"paths":{"/api/health":{"get":{"summary":"Healthcheck","operationId":"healthcheck_api_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiMessage"}}}}}}},"/api/upload":{"post":{"summary":"Upload Reference","operationId":"upload_reference_api_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_reference_api_upload_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/clone":{"post":{"summary":"Create Speaker","operationId":"create_speaker_api_clone_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeakerCreateRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeakerResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/speakers":{"get":{"summary":"List Speakers","operationId":"list_speakers_api_speakers_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/SpeakerResponse"},"type":"array","title":"Response List Speakers Api Speakers Get"}}}}}}},"/api/speakers/{speaker_id}":{"patch":{"summary":"Update Speaker","operationId":"update_speaker_api_speakers__speaker_id__patch","parameters":[{"name":"speaker_id","in":"path","required":true,"schema":{"type":"string","title":"Speaker Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeakerUpdateRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpeakerResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"summary":"Delete Speaker","operationId":"delete_speaker_api_speakers__speaker_id__delete","parameters":[{"name":"speaker_id","in":"path","required":true,"schema":{"type":"string","title":"Speaker Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/generate":{"post":{"summary":"Generate Audio","operationId":"generate_audio_api_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerationRequest"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/generate/stream/{generation_id}":{"get":{"summary":"Stream Generation","operationId":"stream_generation_api_generate_stream__generation_id__get","parameters":[{"name":"generation_id","in":"path","required":true,"schema":{"type":"string","title":"Generation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/history":{"get":{"summary":"Get History","operationId":"get_history_api_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/GenerationResponse"},"type":"array","title":"Response Get History Api History Get"}}}}}}},"/api/history/{generation_id}/download":{"get":{"summary":"Download History Audio","operationId":"download_history_audio_api_history__generation_id__download_get","parameters":[{"name":"generation_id","in":"path","required":true,"schema":{"type":"string","title":"Generation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/history/{generation_id}/regenerate":{"post":{"summary":"Regenerate History","operationId":"regenerate_history_api_history__generation_id__regenerate_post","parameters":[{"name":"generation_id","in":"path","required":true,"schema":{"type":"string","title":"Generation Id"}}],"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerationResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/history/{generation_id}":{"delete":{"summary":"Delete History","operationId":"delete_history_api_history__generation_id__delete","parameters":[{"name":"generation_id","in":"path","required":true,"schema":{"type":"string","title":"Generation Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiMessage"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"ApiMessage":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"ApiMessage"},"Body_upload_reference_api_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_reference_api_upload_post"},"GenerationRequest":{"properties":{"speaker_id":{"type":"string","title":"Speaker Id"},"text":{"type":"string","maxLength":2000,"minLength":1,"title":"Text"},"language":{"type":"string","pattern":"^(zh|en|ja|ko|fr|de|es)$","title":"Language"},"voice_similarity":{"type":"integer","maximum":100.0,"minimum":0.0,"title":"Voice Similarity","default":80},"emotion":{"type":"integer","maximum":100.0,"minimum":0.0,"title":"Emotion","default":50},"emotion_label":{"type":"string","pattern":"^(Happy|Sad|Excited|Calm|Serious)$","title":"Emotion Label","default":"Calm"},"speed":{"type":"number","maximum":2.0,"minimum":0.5,"title":"Speed","default":1.0},"pitch":{"type":"integer","maximum":12.0,"minimum":-12.0,"title":"Pitch","default":0},"stability":{"type":"integer","maximum":100.0,"minimum":0.0,"title":"Stability","default":75},"output_format":{"type":"string","pattern":"^(wav)$","title":"Output Format","default":"wav"}},"type":"object","required":["speaker_id","text","language"],"title":"GenerationRequest"},"GenerationResponse":{"properties":{"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"id":{"type":"string","title":"Id"},"speaker":{"$ref":"#/components/schemas/SpeakerResponse"},"text":{"type":"string","title":"Text"},"language":{"type":"string","title":"Language"},"voice_similarity":{"type":"integer","title":"Voice Similarity"},"emotion":{"type":"integer","title":"Emotion"},"emotion_label":{"type":"string","title":"Emotion Label"},"speed":{"type":"number","title":"Speed"},"pitch":{"type":"integer","title":"Pitch"},"stability":{"type":"integer","title":"Stability"},"output_format":{"type":"string","title":"Output Format"},"engine":{"type":"string","title":"Engine"},"status":{"type":"string","title":"Status"},"audio_url":{"type":"string","title":"Audio Url"},"meta":{"type":"object","title":"Meta"}},"type":"object","required":["created_at","updated_at","id","speaker","text","language","voice_similarity","emotion","emotion_label","speed","pitch","stability","output_format","engine","status","audio_url","meta"],"title":"GenerationResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"SpeakerCreateRequest":{"properties":{"upload_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Upload Id"},"name":{"type":"string","maxLength":255,"minLength":2,"title":"Name"},"favorite":{"type":"boolean","title":"Favorite","default":false},"base_speaker_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Base Speaker Id"},"mix_with_speaker_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mix With Speaker Id"},"mix_ratio_primary":{"type":"integer","maximum":100.0,"minimum":0.0,"title":"Mix Ratio Primary","default":70},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["name"],"title":"SpeakerCreateRequest"},"SpeakerResponse":{"properties":{"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"id":{"type":"string","title":"Id"},"name":{"type":"string","title":"Name"},"favorite":{"type":"boolean","title":"Favorite"},"upload_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Upload Id"},"engine":{"type":"string","title":"Engine"},"voice_signature":{"type":"object","title":"Voice Signature"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","required":["created_at","updated_at","id","name","favorite","upload_id","engine","voice_signature","notes"],"title":"SpeakerResponse"},"SpeakerUpdateRequest":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":255,"minLength":2},{"type":"null"}],"title":"Name"},"favorite":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Favorite"},"notes":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notes"}},"type":"object","title":"SpeakerUpdateRequest"},"UploadResponse":{"properties":{"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"},"id":{"type":"string","title":"Id"},"original_filename":{"type":"string","title":"Original Filename"},"content_type":{"type":"string","title":"Content Type"},"size_bytes":{"type":"integer","title":"Size Bytes"},"duration_seconds":{"type":"number","title":"Duration Seconds"},"sample_rate":{"type":"integer","title":"Sample Rate"},"channels":{"type":"integer","title":"Channels"},"denoised":{"type":"boolean","title":"Denoised"},"vad_segments":{"items":{"type":"object"},"type":"array","title":"Vad Segments"},"speaker_features":{"type":"object","title":"Speaker Features"},"detected_language":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Detected Language"},"prompt_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Prompt Text"},"normalized_media_url":{"type":"string","title":"Normalized Media Url"}},"type":"object","required":["created_at","updated_at","id","original_filename","content_type","size_bytes","duration_seconds","sample_rate","channels","denoised","vad_segments","speaker_features","detected_language","prompt_text","normalized_media_url"],"title":"UploadResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}